Update README.md
Moved the old wiki page I'd written into the readme. Added the required licensing per Windower to the readme. We shouldn't have any sort of additional licenses we have to include with any files, as we're just going to be submitting .dbs to the main repo. All CYA should be done now.